+++ contracts/interfaces/Velodrome/IOptiSwapPair.sol
@@ -2,5 +2,10 @@
 pragma solidity 0.8.17;
 
 interface IOptiSwapPair {
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
+    function swap(
+        uint256 amount0Out,
+        uint256 amount1Out,
+        address to,
+        bytes calldata data
+    ) external;
 }
