+++ contracts/interfaces/Velodrome/IVeloGauge.sol
@@ -2,19 +2,22 @@
 pragma solidity 0.8.17;
 
 interface IVeloGauge {
-    function notifyRewardAmount(address token, uint amount) external;
+    function notifyRewardAmount(address token, uint256 amount) external;
 
     function getReward(address account, address[] calldata tokens) external;
 
-    function claimFees() external returns (uint claimed0, uint claimed1);
+    function claimFees() external returns (uint256 claimed0, uint256 claimed1);
 
-    function left(address token) external view returns (uint);
+    function left(address token) external view returns (uint256);
 
     function isForPair() external view returns (bool);
 
-    function earned(address token, address account) external view returns (uint);
+    function earned(address token, address account)
+        external
+        view
+        returns (uint256);
 
-    function balanceOf(address account) external view returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 
     function deposit(uint256 amount, uint256 tokenId) external;
