+++ contracts/layerzero/OFTV2.sol
@@ -6,24 +6,36 @@ import "../token/ERC20.sol";
 import "./BaseOFTV2.sol";
 
 abstract contract OFTV2 is BaseOFTV2, ERC20 {
-
-    uint internal immutable ld2sdRate;
+    uint256 internal immutable ld2sdRate;
 
     constructor(
-        string memory _name, 
-        string memory _symbol, 
-        uint8 _sharedDecimals, 
-        address _lzEndpoint, 
-        ICentralRegistry _centralRegistry) ERC20(_name, _symbol) BaseOFTV2(_sharedDecimals, _lzEndpoint, _centralRegistry) {
+        string memory _name,
+        string memory _symbol,
+        uint8 _sharedDecimals,
+        address _lzEndpoint,
+        ICentralRegistry _centralRegistry
+    )
+        ERC20(_name, _symbol)
+        BaseOFTV2(_sharedDecimals, _lzEndpoint, _centralRegistry)
+    {
         uint8 decimals = decimals();
-        require(_sharedDecimals <= decimals, "OFT: sharedDecimals must be <= decimals");
-        ld2sdRate = 10 ** (decimals - _sharedDecimals);
+        require(
+            _sharedDecimals <= decimals,
+            "OFT: sharedDecimals must be <= decimals"
+        );
+        ld2sdRate = 10**(decimals - _sharedDecimals);
     }
 
     /************************************************************************
-    * public functions
-    ************************************************************************/
-    function circulatingSupply() public view virtual override returns (uint) {
+     * public functions
+     ************************************************************************/
+    function circulatingSupply()
+        public
+        view
+        virtual
+        override
+        returns (uint256)
+    {
         return totalSupply();
     }
 
@@ -32,29 +44,43 @@ abstract contract OFTV2 is BaseOFTV2, ERC20 {
     }
 
     /************************************************************************
-    * internal functions
-    ************************************************************************/
-    function _debitFrom(address _from, uint16, bytes32, uint _amount) internal virtual override returns (uint) {
+     * internal functions
+     ************************************************************************/
+    function _debitFrom(
+        address _from,
+        uint16,
+        bytes32,
+        uint256 _amount
+    ) internal virtual override returns (uint256) {
         address spender = _msgSender();
         if (_from != spender) _spendAllowance(_from, spender, _amount);
         _burn(_from, _amount);
         return _amount;
     }
 
-    function _creditTo(uint16, address _toAddress, uint _amount) internal virtual override returns (uint) {
+    function _creditTo(
+        uint16,
+        address _toAddress,
+        uint256 _amount
+    ) internal virtual override returns (uint256) {
         _mint(_toAddress, _amount);
         return _amount;
     }
 
-    function _transferFrom(address _from, address _to, uint _amount) internal virtual override returns (uint) {
+    function _transferFrom(
+        address _from,
+        address _to,
+        uint256 _amount
+    ) internal virtual override returns (uint256) {
         address spender = _msgSender();
         // if transfer from this contract, no need to check allowance
-        if (_from != address(this) && _from != spender) _spendAllowance(_from, spender, _amount);
+        if (_from != address(this) && _from != spender)
+            _spendAllowance(_from, spender, _amount);
         _transfer(_from, _to, _amount);
         return _amount;
     }
 
-    function _ld2sdRate() internal view virtual override returns (uint) {
+    function _ld2sdRate() internal view virtual override returns (uint256) {
         return ld2sdRate;
     }
 }
